requirements.txt
langchain==0.0.267
openai==0.27.0
chromadb==0.4.13
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.10
streamlit==1.26.0
pandas==2.0.3
matplotlib==3.7.2
nltk==3.8.1
python-dotenv==1.0.0
lxml==4.9.3
tqdm==4.66.1